Added connectivity, connection_controller, connection status, bindings.

+++ lib/services/database.dart
@@ -17,11 +17,18 @@ class DatabaseService {
 
   ///Adds weight to the database.
   Future<void> addWeight(
+    String connection,
     CollectionReference<Object?> weights,
     TextEditingController textEditingController,
   ) async {
     try {
-      if (double.parse(textEditingController.text) < 1.0) {
+      if(connection != 'mobile' && connection != 'wifi') {
+        throw FirebaseException(
+          plugin: 'firebase',
+          code: 'network-error',
+          message: 'Weight cannot be added. \nCheck your Internet connection.');
+      } else
+        if (double.parse(textEditingController.text) < 1.0) {
         throw Exception('Weight must be at least 1.');
       }
 
@@ -40,8 +47,15 @@ class DatabaseService {
       Get.snackbar(
         'Error',
         'Weight cannot be added. \nCheck your Internet connection.',
+        backgroundColor:Colors.orange,
+        colorText: Colors.white,
+        borderColor: Colors.white,
+        borderWidth: 1.0,
         snackPosition: SnackPosition.BOTTOM,
-        icon: const Icon(Icons.error),
+        icon: const Icon(
+          Icons.error,
+          color: Colors.white,
+        ),
         margin: const EdgeInsets.all(8.0),
       );
       await Future.delayed(const Duration(seconds: 4));
@@ -49,8 +63,15 @@ class DatabaseService {
       Get.snackbar(
         'Invalid weight',
         'Weight cannot be added due to invalid number.',
+        backgroundColor:Colors.orange,
+        colorText: Colors.white,
+        borderColor: Colors.white,
+        borderWidth: 1.0,
         snackPosition: SnackPosition.BOTTOM,
-        icon: const Icon(Icons.error),
+        icon: const Icon(
+          Icons.error,
+          color: Colors.white,
+        ),
         margin: const EdgeInsets.all(8.0),
       );
       await Future.delayed(const Duration(seconds: 4));
@@ -59,10 +80,17 @@ class DatabaseService {
 
   ///Updates weight in the database.
   void updateWeight(
+    String connection,
     dynamic weight,
     TextEditingController textEditingController,
   ) async {
     try {
+      if(connection != 'mobile' && connection != 'wifi') {
+        throw FirebaseException(
+            plugin: 'firebase',
+            code: 'network-error',
+            message: 'Weight cannot be updated. \nCheck your Internet connection.');
+      } else
       if (double.parse(textEditingController.text) < 1.0) {
         throw Exception('Weight must be at least 1.');
       }
@@ -75,27 +103,63 @@ class DatabaseService {
       Get.back();
       Get.snackbar(
         'Error',
-        'Weight cannot be edited.',
-        icon: const Icon(Icons.error),
+        'Weight cannot be updated. \nCheck your Internet connection.',
+        backgroundColor:Colors.orange,
+        colorText: Colors.white,
+        borderColor: Colors.white,
+        borderWidth: 1.0,
         snackPosition: SnackPosition.BOTTOM,
+        icon: const Icon(
+          Icons.error,
+          color: Colors.white,
+        ),
         margin: const EdgeInsets.all(8.0),
       );
     } on Exception catch (e) {
       Get.back();
       Get.snackbar(
         'Invalid weight',
-        'Weight cannot be edited due to invalid number.',
-        snackPosition:SnackPosition.BOTTOM,
+        'Weight cannot be updated due to invalid number.',
         backgroundColor:Colors.orange,
         colorText: Colors.white,
-        icon: const Icon(Icons.error),
+        borderColor: Colors.white,
+        borderWidth: 1.0,
+        snackPosition:SnackPosition.BOTTOM,
+        icon: const Icon(
+          Icons.error,
+          color: Colors.white,
+        ),
         margin: const EdgeInsets.all(8.0),
       );
     }
   }
 
   ///Deletes weight from the database.
-  void deleteWeight(dynamic weight) {
-    weight.reference.delete();
+  void deleteWeight(String connection, dynamic weight) {
+    try {
+      if(connection != 'mobile' && connection != 'wifi') {
+        throw FirebaseException(
+          plugin: 'firebase',
+          code: 'network-error',
+          message: 'Weight cannot be deleted. \nCheck your Internet connection.');
+      }
+
+      weight.reference.delete();
+    } on FirebaseException catch (e) {
+      Get.snackbar(
+        'Error',
+        'Weight cannot be deleted. \nCheck your Internet connection.',
+        backgroundColor:Colors.orange,
+        colorText: Colors.white,
+        borderColor: Colors.white,
+        borderWidth: 1.0,
+        snackPosition:SnackPosition.BOTTOM,
+        icon: const Icon(
+          Icons.error,
+          color: Colors.white,
+        ),
+        margin: const EdgeInsets.all(8.0),
+      );
+    }
   }
 }
